requirements.txt
wheel==0.29.0
pytz==2017.2
boto3==1.7.76
mysql-connector-python-rf==2.2.2
PyYAML==3.12
minio==2.2.4
#cassandra-driver==3.12.0
kafka==1.3.5
influxdb==5.0.0
pyarrow==0.8.0
pympler==0.5
hdfs3==0.3.0